# server/helpers/swagger.yaml
basePath: /
consumes:
- application/json
- multipart/form-data
definitions:
...
  Map:
    additionalProperties:
      type: object
    description: |-
      It implements native SQL driver interfaces and hence can
      be used for SQL json or jsonb types as a drop in replacement
      of golang native maps
    title: Map type is an alias for map[string]interface{}
    type: object
    x-go-package: github.com/layer5io/meshery/server/internal/sql
...
  MesheryPattern:
    description: MesheryPattern represents the patterns that needs to be saved
    properties:
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      id:
        $ref: '#/definitions/UUID'
      location:
        $ref: '#/definitions/Map'
      name:
        type: string
        x-go-name: Name
      pattern_file:
        type: string
        x-go-name: PatternFile
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
      user_id:
        description: |-
          Meshery doesn't have the user id fields
          but the remote provider is allowed to provide one
        type: string
        x-go-name: UserID
    type: object
    x-go-package: github.com/layer5io/meshery/server/models
...
  UUID:
    items:
      format: uint8
      type: integer
    title: UUID is an array type to represent the value of a UUID, as defined in RFC-4122.
    type: array
    x-go-package: github.com/google/uuid
...
info:
  description: |-
    the purpose of this application is to provide an application
    that is using plain go code to define an API

    This should demonstrate all the possible comment annotations
    that are available to turn go code into a fully compliant swagger 2.0 spec
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.txt
  title: Meshery API.
  version: 0.4.27
paths:
...
  /api/pattern/{id}:
    delete:
      description: DeleteMultiMesheryPatternsHandler deletes patterns with the given
        ids
      operationId: idDeleteMesheryPattern
      parameters:
      - description: id for a specific
        format: uuid
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/noContentWrapper'
      summary: Handle Delete for multiple Meshery Patterns
      tags:
      - PatternsAPI
    get:
      description: Fetches the pattern with the given id
      operationId: idGetMesheryPattern
      parameters:
      - description: id for a specific
        format: uuid
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/mesheryPatternResponseWrapper'
      summary: Handle GET for a Meshery Pattern
      tags:
      - PatternsAPI
...
  /api/patterns:
    delete:
      description: DeleteMultiMesheryPatternsHandler deletes patterns with the given
        ids
      operationId: idDeleteMesheryPattern
      parameters:
      - description: id for a specific
        format: uuid
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/noContentWrapper'
      summary: Handle Delete for multiple Meshery Patterns
      tags:
      - PatternsAPI
...
produces:
- application/json
responses:
...
  mesheryPatternResponseWrapper:
    description: Returns a single meshery pattern
    schema:
      $ref: '#/definitions/MesheryPattern'
...
  noContentWrapper:
    description: ""
...
schemes:
- http
security:
- token:
  - '[]'
securityDefinitions:
  token:
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"